🎉 feat(event-n-this.html): add HTML file for Events & The Keyword This notes This commit adds an HTML file for the Events & The Keyword This notes. The file contains a series of buttons and headings that will be used to demonstrate the behavior of the "this" keyword in JavaScript events.
🎨 style(event-n-this.js): refactor code to use "this" keyword in event listener
Refactored the code to use the "this" keyword in the event listener to select the element being clicked on or whatever event is being listened to. This eliminates duplicate code when the only thing that is changing is the selected element. The colorize function was created to handle the styling logic, and it is now called instead of an anonymous function after the click event.

✨ feat(keyboard-events): add HTML and JavaScript files for keyboard events and event objects
This commit adds an HTML file and a JavaScript file for keyboard events and event objects. The HTML file contains the basic structure of the page, while the JavaScript file will contain the code for handling keyboard events and event objects.