✨ feat(event-delegation.html): add HTML file for Event Delegation notes This commit adds an HTML file for the Event Delegation notes. The file contains a form for posting tweets and a list of tweets. The list contains two li elements and a p element that is not an li element.
📝 docs(event-delegation.js): add event delegation to tweet form and tweets container
This commit adds event delegation to the tweet form and tweets container. The tweet form now listens for a submit event and prevents the default form submit behaviour. The tweets container listens for a click event and removes the clicked list item. This improves the functionality of the application as it allows for newly created list items to have event listeners.