🔧 chore(.eslintrc.json): turn off no-unused-expressions rule 📝 docs(callback-hell.html): add HTML file for Callback Hell notes The no-unused-expressions rule is turned off in the .eslintrc.json file to allow for unused expressions in the code. An HTML file is added for the Callback Hell notes.
📝 docs(callback-hell.js): add comments and refactor code
This commit adds comments to the code to explain what is happening and how the code works. It also refactors the code to use a function that takes in the new color, delay, and a callback function to execute after the color change. This function is then called with the desired colors and delays, and the callbacks are nested to create a rainbow effect. The commit also mentions the drawbacks of using callbacks and how promises and async functions can help alleviate the issue.