🎨 style(promises.html): add background color transition to body This commit adds a background color transition to the body of the HTML document. The transition will take 0.5 seconds to complete.
🔥 chore(promises.js): comment out unused code
The code that defines the `fakeRequestPromise` function is commented out as it is not being used in the current implementation.

🔥 refactor(promises.js): refactor callback functions to use Promises
The code has been refactored to use Promises instead of callback functions. The fakeRequestPromise function has been replaced with a new fakeRequest function that returns a Promise. The code has been updated to use the Promise.then() and Promise.catch() methods to handle the asynchronous requests. The rainbowColor function has also been updated to use Promises and the code has been refactored to use implicit returns.