🎨 style(axios.html): add button and ul elements to the page The button and ul elements are added to the page to allow users to click and get new jokes. The ul element will be used to append the jokes to when the button is clicked.
📝 docs(axios.js): add comments to explain headers and axios usage
This commit adds comments to explain how to use headers with axios. It also adds a function that requests a new joke from an API and appends it to an unordered list element in the HTML document. The function uses axios to make a GET request to the API and passes in a headers object as a configuration object in the second argument of the axios.get() method. The headers object contains the Accept key with a value of "application/json". The function then returns the joke text from the API response. The addNewJoke() function is an event listener for a button click and calls the getDadJoke() function to get a new joke and appends it to the unordered list element.