📝 docs(prototypes): add notes on prototypes and arrays Added notes on prototypes and arrays. Arrays are objects in JS and they have several methods. Prototypes are template objects that contain a bunch of methods, and we can create multiple objects that share the same prototype so they all have access to the same methods without having to have their own copy. The `__proto__` references the array prototype. Each type of object has a prototype object which is a part of every object of a type. You can even create your own custom object and set it to an object's prototype such as an array's prototype.
🎉 feat(prototype.js, oop.html): add prototype and OOP notes
The prototype.js file contains examples of how to add custom methods to the String and Array prototypes. The code demonstrates how to add a custom method to the String prototype and override an existing method on the Array prototype. The oop.html file is a new file that contains a basic HTML template for notes on Object Oriented Programming.

📝 docs(oop.js): add introductory comments on Object Oriented Concepts
This commit adds introductory comments on Object Oriented Concepts. It highlights the different ways of organizing, designing, and structuring applications by breaking things up into distinct patterns/recipes of objects. The examples given include Factory Functions, Constructor Functions, new CLASS keyword, constructors in a class, the NEW keyword, super and inheritance.