🚀 feat(classes.js): add hsl, fullSat and opposite methods to Color class The hsl method returns the color in HSL format, fullSat method returns the color with full saturation, and opposite method returns the opposite color on the color wheel by adding 180 to the hue value. These methods are added to the Color class and can be used to manipulate colors.
🚀 feat(classes.js): add calcHSL method to Color class
The calcHSL method is added to the Color class, which converts an RGB color to HSL. The method takes the RGB values of the color and calculates the corresponding HSL values. The HSL values are then assigned to the object. A new color, orange, is also defined using the Color class.

🔍 docs(classes.js): add comments and examples for new HSL methods
This commit adds comments and examples for the new HSL methods added to the Color class. The `calcHSL()` method calculates the HSL values for a given color and adds them to the Color object. The `hsl()` method returns the HSL values for a given color. The `opposite()` method returns the opposite color of a given color. The `fullSat()` method returns the fully saturated version of a given color. Examples are provided to demonstrate the usage of these methods.