🎉 feat(index.js): add support for EJS templating engine This commit adds support for the EJS templating engine to the Express application. The app.set() method is used to set the view engine to EJS and the views directory to the /views folder. The app.get() method is used to define two routes, one for the home page and one for a random number page. The res.render() method is used to render the EJS templates and pass data to them.
🎉 feat(ejs, package.json): add ejs dependency and create two new views
The ejs dependency was added to the project to allow for the creation of dynamic HTML pages. Two new views were created, home.ejs and random.ejs. The home.ejs view contains a basic HTML structure with some EJS code to demonstrate how to use EJS. The random.ejs view is a simple HTML page that displays a random number passed in from the index.js file.