🎨 style(index.js): add routes for cats and subreddit 🆕 feat(cats.ejs): add view for all cats Added two new routes to the application, one for displaying a list of cats and another for displaying a subreddit-like page. The cats route uses an array of cat names to simulate data from a database and passes it to the cats.ejs view. The cats.ejs view is a new file that displays all the cats in the array.
🎨 style(random.ejs): add logic to display whether the random number is even or odd
🆕 feat(subreddit.ejs): add a new view for browsing subreddits
The random.ejs file now displays whether the random number generated is even or odd. This is done by adding logic to the ejs file to check if the number is even or odd and display the appropriate message. The subreddit.ejs file is a new view that displays the name of the subreddit being browsed. It is a simple file that only contains an h1 tag with the name of the subreddit.