📝 chore(data.json): add subreddit data This commit adds a new file, data.json, which contains data for three subreddits: soccer, chickens, and mightyharvest. Each subreddit has a name, number of subscribers, description, and an array of posts. The posts have a title, author, and an optional image. This data will be used in the subreddit-express exercise.
🎉 feat(subreddit-express): add subreddit route and cats view
This commit adds a new route to the express application that handles subreddit requests. The route uses the subreddit name as a parameter to retrieve data from a JSON file and pass it to a template. If the subreddit does not exist in the JSON file, a notfound template is rendered. Additionally, a new view for cats is added that displays a list of all the cats in an array.

✨ feat(views): add home.ejs and notfound.ejs templates
Two new templates have been added to the views directory. The home.ejs template contains basic HTML structure with some sample text and an example of using EJS to render dynamic content. The notfound.ejs template contains a simple message that informs the user that the requested subreddit could not be found.

🎨 style(random.ejs): add HTML comments and improve readability
This commit adds HTML comments to the random.ejs file to improve readability and make it easier to understand the purpose of each section of code. Additionally, some code has been refactored to improve readability, such as removing unnecessary comments and using ternary operators to make the code more concise.

🆕 feat(subreddit.ejs): add subreddit view template
This commit adds a new view template for the subreddit page. The template includes HTML and EJS syntax to render the subreddit name, description, and number of subscribers. It also includes a loop to render all posts in the subreddit. If a post has an image, it will be displayed along with the post title and author.